Darwin: mirror specified folders to another location on local drive.
The destination is treated as a read-only mirror of the source. This is useful
as a way to put the contents of Syncthing-sync'ed folders into Dropbox. (Here,
Dropbox isn't a backup, but it is a very convenient tool for accessing per-file
version history, as well as enabling file access on mobile.)